pyproject.toml
[tool.poetry]
name = "atlantic_server"
version = "0.3.0"
description = "Server side of an application of an Aircraft Technical Log"
authors = ["Matthieu Nué <matthieu.nue@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2.6"
djangorestframework = "^3.10.3"
drf-nested-routers = "^0.91"
django-filter = "^2.2.0"
pygit2 = "~1.0"
cffi = "^1.14.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3"
pylint = "^2.4.3"
black = {version = "^19.3b0", allow-prereleases = true}
pylint-django = "^2.0.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
